backend/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.5
django-cookies-samesite==0.9.0
django-cors-headers==3.11.0
django-dcors==0.1
django-elasticsearch-dsl==7.2.1
django-elasticsearch-dsl-drf==0.22.4
django-nine==0.2.5
djangorestframework==3.13.1
elasticsearch==7.16.2
elasticsearch-dsl==7.4.0
idna==3.3
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2021.3
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
ua-parser==0.10.0
urllib3==1.26.8
